+++ src/main/java/org/pio/ui/sidePanel/aSidePanel.java
@@ -6,11 +6,11 @@ import java.awt.*;
 import java.util.LinkedHashMap;
 
 public abstract class aSidePanel implements sidePanelMethods {
-    protected int width, height, posX, posY;
+    protected float width, height, posX, posY;
     public LinkedHashMap<Integer, bRectangle> currentDataLinkedMap;
     public LinkedHashMap<Integer, bRectangle> allDataLinkedMap;
     protected Shape sidePanelBounds;
-    public aSidePanel(int width, int height, int posX, int posY) {
+    public aSidePanel(float width, float height, float posX, float posY) {
         this.width = width;
         this.height = height;
         this.posX = posX;
@@ -20,7 +20,7 @@ public abstract class aSidePanel implements sidePanelMethods {
     }
 
     private Shape initRectangleBounds() {
-        return new Rectangle(posX, posY, width, height);
+        return new Rectangle((int) posX, (int) posY, (int) width, (int) height);
     }
     public Shape getSidePanelBounds() {
         return sidePanelBounds;
@@ -33,7 +33,7 @@ public abstract class aSidePanel implements sidePanelMethods {
     }
     private void drawBody(Graphics g) {
         g.setColor(Color.black);
-        g.fillRect(posX, posY, width, height);
+        g.fillRect((int) posX, (int) posY, (int) width, (int) height);
     }
     protected void drawButtons(Graphics g) {
